{{cookiecutter.project_name}}/requirements/base.txt
Django==1.10.6
django-environ==0.4.1
djangorestframework==3.6
unipath==1.1
elasticsearch==5.2.0
Markdown==2.6.8
Sphinx==1.5.3
sphinx-rtd-theme==0.2.4
django-rest-swagger==2.1.2
pypandoc==1.3.3
arrow==0.10.0
gunicorn==19.7.1
whitenoise==3.3.0
{% if cookiecutter.use_logstash == 'y' -%}
python-logstash==0.4.6
{% endif %}